### 2021-07-12 #3468 @jbytecode
Hi @qpmnguyen and @rchg, would you be willing to assist in reviewing this submission for JOSS?

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "NoisySignalIntegration.jl: A Julia package for uncertainty evaluation of numeric integrals". You can find more information at the top of this Github issue.
