### 2021-07-22 #3468 @jbytecode
Hi @mseri 👋👋👋

Would you be willing to assist in reviewing this submission for JOSS (Journal of Open Source Software) ?

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "NoisySignalIntegration.jl: A Julia package for uncertainty evaluation of numeric integrals". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/3468).
