### 2021-07-23 #3526 @whedon
[REVIEW]: NoisySignalIntegration.jl: A Julia package for uncertainty evaluation of numeric integrals
**Submitting author:** @nluetts (<a href="http://orcid.org/0000-0001-8459-1714">Nils Lüttschwager</a>)
**Repository:** <a href="https://github.com/nluetts/NoisySignalIntegration.jl" target ="_blank">https://github.com/nluetts/NoisySignalIntegration.jl</a>
**Version:** v0.2.2
