+++ paper/paper.bib
@@ -25,20 +25,18 @@ eprint = {https://doi.org/10.1063/1.5009011}
 }
 
 @article{distributions1,
-    author = {{Besan{\c{c}}on}, Mathieu and {Anthoff}, David and {Arslan}, Alex and
-    {Byrne}, Simon and {Lin}, Dahua and {Papamarkou}, Theodore and
-    {Pearson}, John},
-    title = {Distributions.jl: Definition and Modeling of Probability Distributions in the JuliaStats Ecosystem},
-    keywords = {Statistics - Computation, Computer Science - Mathematical Software},
-    year = 2019,
-    month = "Jul",
-    eid = {arXiv:1907.08611},
-    pages = {arXiv:1907.08611},
-    eprint = {1907.08611},
-    primaryClass = {stat.CO},
-    url = {https://arxiv.org/abs/1907.08611}
+   author = {Mathieu Besançon and Theodore Papamarkou and David Anthoff and Alex Arslan and Simon Byrne and Dahua Lin and John Pearson},
+   title = {Distributions.jl: Definition and Modeling of Probability Distributions in the JuliaStats Ecosystem},
+   journal = {Journal of Statistical Software},
+   volume = {98},
+   number = {16},
+   year = {2021},
+   keywords = {Julia; distributions; modeling; interface; mixture; KDE; sampling; probabilistic programming; inference},
+   issn = {1548-7660},
+   pages = {1--30},
+   doi = {10.18637/jss.v098.i16},
+   url = {https://www.jstatsoft.org/v098/i16}
 }
-
 % reference for the software itself
 @misc{distributions2,
   author       = {Dahua Lin and
@@ -102,8 +100,19 @@ eprint = {https://doi.org/10.1063/1.5009011}
 @article{gawrilow,
   author = "Maxim Gawrilow and Martin A. Suhm",
   title = "Quantifying Conformational Isomerism in Chain Molecules by Linear {R}aman Spectroscopy: The Case of Methyl Esters",
-  journal = "Preprints",
+  journal = "Molecules",
   year = 2021,
-  pages = 2021070033,
-  doi = "10.20944/preprints202107.0033.v1"
+  volume = 26,
+  pages = 4523,
+  doi = "10.3390/molecules26154523",
+}
+
+@article{zimmermann,
+  author = "Charlotte Zimmermann and Manuel Lange and Martin A. Suhm",
+  title = "Halogens in Acetophenones Direct the Hydrogen Bond Docking Preference of Phenol via Stacking Interactions",
+  journal = "Molecules",
+  year = 2021,
+  volume = 26,
+  pages = 4883,
+  doi = "10.3390/molecules26164883",
 }
