+++ v0.2.2/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"internals/#Internals","page":"Internals","title":"Internals","text":"","category":"section"},{"location":"internals/","page":"Internals","title":"Internals","text":"Upon creation, the data types UncertainCurve and UncertainBound use the package MonteCarloMeasurements.jl to generate random samples. Furthermore, integration results are returned as Particle objects from MonteCarloMeasurements which enables you to use MonteCarloMeasurements for further error propagation. Check out the MonteCarloMeasurements.jl documentation to learn more about the sampling process and uncertainty calculations.","category":"page"},{"location":"guide/#Usage-Guide","page":"Usage Guide","title":"Usage Guide","text":"","category":"section"},{"location":"guide/","page":"Usage Guide","title":"Usage Guide","text":"As a more detailed usage example, we will go through the analysis of a simulated FTIR spectrum.","category":"page"},{"location":"guide/","page":"Usage Guide","title":"Usage Guide","text":"info: A note on plotting\nNoisySignalIntegration provides several \"recipes\" for Plots.jl to easily plot the various (interim) results. Often, merely calling plot() and passing in data types from NoisySignalIntegration will work. Examples are included in this guide.","category":"page"},{"location":"guide/","page":"Usage Guide","title":"Usage Guide","text":"Suppose our spectrum looks like the following simulation:","category":"page"},{"location":"guide/","page":"Usage Guide","title":"Usage Guide","text":"using Distributions, NoisySignalIntegration, Plots\nusing Random: seed!\n\nspectrum = NoisySignalIntegration.testdata_1()\nplot(spectrum, label=\"simulated spectrum\")","category":"page"},{"location":"guide/","page":"Usage Guide","title":"Usage Guide","text":"In order two apply the NoisySignalIntegration uncertainty analysis, we must perform 4 basic steps:","category":"page"},{"location":"guide/","page":"Usage Guide","title":"Usage Guide","text":"From the spectrum, crop the region that contains the signals and the region that contains a representative sample of the noise\nCharacterize the noise (to be able to simulate it in the Monte-Carlo draws) \nSet integration bounds and their associated uncertainties\nRun the mc_integrate function","category":"page"},{"location":"guide/#crop","page":"Usage Guide","title":"Cropping the spectrum","text":"","category":"section"},{"location":"guide/","page":"Usage Guide","title":"Usage Guide","text":"Let's start by dividing the spectrum into the bands we want to integrate and the noise we want to analyse. We can do this by using the crop function.","category":"page"},{"location":"guide/","page":"Usage Guide","title":"Usage Guide","text":"slice_bands = crop(spectrum,  5.0,  40.0)\nslice_noise = crop(spectrum, 40.0, 100.0)\n\nplot(slice_bands; label=\"bands\")\nplot!(slice_noise; label=\"noise\")","category":"page"},{"location":"guide/#Noise-analysis","page":"Usage Guide","title":"Noise analysis","text":"","category":"section"},{"location":"guide/","page":"Usage Guide","title":"Usage Guide","text":"The spectrum has a quite considerable baseline which constitutes a problem when analysing the noise. To prepare the noise spectrum slice_noise for analysis, we create a NoiseSample object. Upon construction of the NoiseSample object, a polynomial is fitted and subtracted to remove the baseline:","category":"page"},{"location":"guide/","page":"Usage Guide","title":"Usage Guide","text":"noise = NoiseSample(slice_noise, 3) # 3 = remove third order polynomial baseline\nnothing # hide","category":"page"},{"location":"guide/","page":"Usage Guide","title":"Usage Guide","text":"Plotting the original slice and the NoiseSample object shows the data after baseline removal:","category":"p
[…]
