+++ paper/paper.bib
@@ -61,7 +61,7 @@ eprint = {https://doi.org/10.1063/1.5009011}
 }
 
 @article{mcm,
-  author    = {Fredrik Bagge Carlson},
+  author    = {Fredrik {Bagge Carlson}},
   title     = {MonteCarloMeasurements.jl: Nonlinear Propagation of Arbitrary Multivariate
                Distributions by means of Method Overloading},
   year = 2020,
@@ -80,7 +80,7 @@ eprint = {https://doi.org/10.1063/1.5009011}
 }
 
 @article{metrolopy,
-  author = "Harold Parks",
+  author = "Harold V. Parks",
   title = "MetroloPy",
   journal = "GitHub Repository",
   url = "https://github.com/nrc-cnrc/MetroloPy/",
@@ -89,8 +89,7 @@ eprint = {https://doi.org/10.1063/1.5009011}
 
 @article{measurements,
   author =	 {{Giordano}, Mos{\`e}},
-  title =	 "{Uncertainty propagation with functionally correlated
-                  quantities}",
+  title =	 "{Uncertainty propagation with functionally correlated quantities}",
   pages =	 {arxiv:1610.08716},
   primaryClass = "physics.data-an",
   year =	 2016,
