### 2021-08-30 231bb4a
try to fix author in footer
