### 2021-08-30 #3526 @jbytecode
@nluetts you can use https://whedon.theoj.org/ for compiling and testing the paper
