### 2021-08-30 #3526 @nluetts
Hi @jbytecode,

I re-generated the paper because in the footer of the page my name is missing an initial.
I tried to fix the problem by putting my name into double quotes, but to no avail.

OK, I will try https://whedon.theoj.org/ for further troubleshooting (I used the docker image so far, but there the name in the footer is not included, just a placeholder).
