### 2021-08-30 #3526 @jbytecode
sure, it is about automatic formatting of your paper's citation info. not a problem
