### 2021-08-30 #3526 @nluetts
The troubleshooting is a bit tricky ...

With docker, the footer says:

> ¿citation_author?, (2021). NoisySignalIntegration.jl: A Julia package for uncertainty evaluation of numeric integrals. Journal of Open Source Software, ¿VOL?(¿ISSUE?), ¿PAGE?. https://doi.org/DOIunavailable

At https://whedon.theoj.org/, I get:

> Mickey Mouse et al., (XXXX). NoisySignalIntegration.jl: A Julia package for uncertainty evaluation of numeric integrals. , X(X), X. https://doi.org/10.21105/joss.0XXXX

Only by requesting the paper here, I get the (close to) correct footer:

> Lüttschwager, N. O., (2021). NoisySignalIntegration.jl: A Julia package for uncertainty evaluation of numeric integrals. Journal of Open Source Software, 6(64), 3526. https://doi.org/10.21105/joss.03526

That should be:

> Lüttschwager, N. O. **B.**, (2021). ...

I could try curly braces like in bibtex format, but I am only guessing ...
