### 2021-08-30 #3526 @nluetts
Curly braces do not compile on the docker image. The [Github action](https://joss.readthedocs.io/en/latest/submitting.html#github-action) gives the same result as docker.

I try one more time here with a html non-breaking space in the author name. Other as the curly braces, this compiles fine.
