### 2021-08-30 #3526 @jbytecode
@nluetts since the shorter citation info allows the first characters of the first and middle names and full of family name, it is correct and i think there is nothing to do with that. if you think it is still a problem, i can discuss this with other editors.
