### 2021-08-30 #3526 @nluetts
@jbytecode if that initial is only missing in the footer and not in the metadata of the paper in general, it is fine.
Otherwise, the draft looks good to me now.

With regard to the other remaining items:

I updated the citation of Distributions.jl and included the DOI.
I also updated the citation of Gawrilow and Suhm, because that paper went from preprint stage to published since I submitted here.
Additionally, I added another reference from my group which uses the package (Zimmermann et al., 2021). I hope that was fine at this late stage.

I made a tagged release, v0.2.2, and archived the code on zenodo.org with DOI 10.5281/zenodo.5338743

https://doi.org/10.5281/zenodo.5338743
