### 2021-12-06 #3526 @kyleniemeyer
@nluetts It is possible to update the paper if something was missing, but it's a bit unusual to add an acknowledgment so long after the fact.

cc @arfon
