### 2021-12-07 #3526 @nluetts
Yes, I admit that this is a bit awkward.

Although I was not funded by it, a research training group which I am associated with was part of the motivation for me to develop the package (they require thorough uncertainty analysis.)

It was pointed out to me that the publication can be included in the research training group's publication list, but this requires that the RTG is mentioned in the acknowledgment of the paper. Therefore, I would like to revise the acknowledgement and add a sentence.

If that is OK, how would I go about it? Can I simply update the paper branch in my repository and you can instruct whedon to rebuild the paper and replace it on the JOSS website?
