### 2021-12-07 #3526 @arfon
> If that is OK, how would I go about it? Can I simply update the paper branch in my repository and you can instruct whedon to rebuild the paper and replace it on the JOSS website?

Yes, I can do this. Please let me know when the paper is updated in your repository.
